SN-7251 - Addressed poor performance in FwPkg/ISBootloader updates. (#991)
* Resolve Windows type-cast warnings, where possible.
* Resolved stalls/hangs when detecting devices in Bootloader during handshaking/sync - ISBFirmwareUpdater, ISDevice, and ISBootloaderISB.
* Improved durability of FirmwareUpdateHost::fwUpdate_processMessage() to session-check only those messages which has a session_id.
* Starting implementation for ISFwUpdaterCmd::resultMsg used to report more detailed status during an update.

+++ src/ISBootloaderThread.h
@@ -54,7 +54,7 @@ public:
         thread_serial_t(const std::string& port_name, bool force_isb_update = false) {
             // FIXME: This is pretty jank... and ridiculous.  I can do better!
             port_handle_t port = (port_handle_t)&(serialPort);
-            serialPortInit(port, m_serial_threads.size(), PORT_TYPE__UART | PORT_TYPE__COMM);
+            serialPortInit(port, (int)m_serial_threads.size(), PORT_TYPE__UART | PORT_TYPE__COMM);
             serialPortPlatformInit(port);
             serialPortSetPort(port, port_name.c_str());
